Mark dependencies of assemblies with the copy and save action
When an assembly had the action Copy, a lot of it's dependencies could go unmarked.  Things like public methods were being marked, but private methods, attributes, events, properties were not.

With this change the MarkStep will visit everything in a copy or saved assembly and mark everything.

I did not address the *Used AssemblyActions.  These will still be vulnerable to invalid IL.  Additional logic would be needed to track the first marking of these assemblies and then call `ForceFullMarkingOfAssembly`.

I did not handle `AddBypassNGen` either.  The XML comment for this action states that `Keep all types, methods, and fields but add System.Runtime.BypassNGenAttribute to unmarked methods.`  I don't know how this action is used,
if it will still strip out EventDefinitions and PropertyDefinitions then additional tweaks would be needed to avoid the calls to `ForceFullMarkingOfProperties` and `ForceFullMarkingOfEvents`.
If someone would like to handle these cases they could do so in a follow on PR.