Fix the illink.sln so that it loads in VS on Win
The trick is apprently adding the Configurations property.
The rest of the changes in the projects is more of a cleanup (moving general property groups up top).

Also renamed the ILLink.Tasks tests to match the folder and the product names.

VS on Windows forced rewrite of the illink.sln file unfortunately - trying to go back to the original ordering doesn't help - VS will overwrite it next time it opens it.
But it's really just order changes and project type GUIDs.