Fix a bug in the reflection pattern recorder usage (#921)
* Fix a bug in the reflection pattern recorder usage

The code was written such that it would report almost all calls it didn't consider (or try to analyze) as unrecognized patterns.

This change adds the test infra to validate that no extra reflection patterns are recorded on a given test. It then uses this on two tests to validate the functionality.

The fix adds a boolean which tracks if the code is trying to analyze a given reflection pattern or not. It only reports unrecognized pattern if it did try to analyze a pattern, but didn't report any result explicitly.

* PR feedback and adding more tests

As per PR feedback added explicit reporting of unrecognized patterns in all branches. The "using" is now asserting if no result was reported (instead of reporting a generic unrecognized message).

Added many tests to improve the coverage of this code. Not all interesting code paths yet.