Ignore missing .ctor or .cctor in PreserveDependency attribute
Currently linker issues a warning if the member referenced in PreserveDependency attribute doesn't exist. This makes perfect sense for most cases, but in case of .ctor and .cctor it's a bit tricky. Default constructor as well as static constructor are often compiler generated and otherwise unrelated changes to the code (or potentially even the compiler) may add or remove these. The PreserveDependency attribute should express intent, in this case dependency on being able to instantiate the target type or be able to use static fields on it. The fact that the compiler/implementation may or may not have the default .ctor/static .cctor is secondary.

This change will skip emitting the warning if the dependency is for default .ctor or static .cctor.

Fixed couple of small problems in the existing tests and added the new cases.