Reference linker via assembly path in tests
This takes advantage of the SDK hook in 3.x which allows overriding
the tasks package via ILLinkTasksPath. This simplifies the integration
tests, which will no longer require a "dotnet pack" before running.

The old targets file, which was used for the packagereference hook, is
being removed because it was fragile and hasn't been kept up-to-date
with SDK changes.

The portable publish scenario is no longer tested since it's not
supported.

This should unblock the arcade SDK update.