Fix pattern matching for Activator.CreateInstance over unrecognized parameter
On .NET Framework there is an overload of Activator.CreateInstance with takes just one parameter but it's not a Type. This should be reported as unrecognized pattern by the linker.

This change also switches in the internal infra to use exceptions instead of asserts as they play nicer with tests.