Don't use nop(s) for removed instructions in RemoveUnreachableBlocksStep
The CLR spec in Backward Branch Contraints recognizes only unconditional
jumps. For any other instruction including nop it mandates

>In particular, if that single-pass analysis arrives at an instruction,
call it location X, that immediately follows an unconditional branch,
and where X is not the target of an earlier branch instruction, then the
state of the evaluation stack at X, clearly, cannot be derived from
existing information. In this case, the CLI demands that the evaluation

Fixes #975