Fix ref assembly symbols issue (try 2)
https://github.com/mono/linker/pull/1029 didn't fix the issue because
Arcade sets its own defaults that override those in
Directory.Build.props. This moves the import so that our setting wins.

Validated by doing a local build, setting OfficialBuildId manually. I
did the same validation in the last change, but maybe I was working
with a dirty build that somehow had a pdb. I think this should
actually fix it.