Make the dataflow json file general purpose
This makes the file format more general purpose and renames the command line argument to signify that.

I didn't attempt to extract the parser out into a reusable class. We can do that when there's actual need to use this for something else - this is just to make sure the outside interface and structure is general purpose.