Allow stripping resources with ignore-descriptors
This makes --strip-resources and --ignore-descriptors independent, by
always running BlacklistStep, but only adding ResolveFromXmlSteps for
embedded resources if IgnoreDescriptors was false. BlacklistStep is now
responsible for determining whether to strip an embedded resource, so that
the stripping can happen regardless of whether the descriptors were
honored.

Embedded substitutions are tied to the same options, so they are honored
and/or stripped under the same circumstances as the embedded
descriptors. Note in particular that --ignore-descriptors will also
ignore embedded substitutions. Maybe it is worth picking a better name
for this argument (--ignore-resources?) or introducing a new one for the
substitutions only (--ignore-substitutions?).