Improve messages for reflection warnings
Adds a location for messages which don't have source code location. For now this is added as the name of the method where the warning occured. This is definitely a temporary improvement and we should do better once we invest more into the location information.

Fixes several smaller issues with warning messages:
- In some cases Cecil doesn't have parameter names, so use parameter index in that case
- Correctly specify the "this" parameter (which doesn't have a Cecil representation) as the method definition itself which will lead to the correct message