Allow types and other metadata item as source for reflection access pattern
With the upcomming support for DynamicallyAccessedMembersAttribute on generic parameters, the source of the access pattern will no longer be any code, but it can be a type declaration or field definition and so on. In such cases trying to point back to some line of code is complicated and in lot of cases confusing.

This change allows more values to be passed as the source of a reflection access pattern. Currently there's no use case for it. The one which comes close is attribute fields, but those make sense to use the attribute ctor as the source of the problem and use the field as the "accessed" item.

Corresponding changes to the test infra to allow for wider range of values and to handle missing values better.