Recognize Type.AssemblyQualifiedName as intrinsic
This will propagate whatever annotations there are on the type to the string.

Makes it possible to recognize the pattern used by ComponentModel/TypeDescriptionProviderAttribute as safe (the pattern is to accept both a `string` and a `Type`, store the value as a `string`, and `Type.GetType` the type out of that string, before passing it to `Activator`).