This allows linker to propagate constant return values higher up the callstack
This was probably a "bug" as the code is designed to do this and it will work for methods with multiple branches, but it doesn't work for methods which simply return a value from another method (no branching).

This has small impact - in CoreLib this actually only kicks in about 4 times.