Rename DefaultConstructor to PublicParameterlessConstructor
This turned out to be important as we used the enum name in linker attribute XML as well (so we would not recognize the correct name).

I tried to rename everything, including test type/method names and so on.

This also includes remaining renames of DynamicallyAccessedMemberKinds to DynamicallyAccessedMemberTypes (mostly local variable names and such).

No functional changes.