Keep base fields of non-auto layout classes
If a field is accessed on a sequential/explicit class, we need to keep the instance fields on base types intact so that the offsets match. This happens naturally, except for the case where the type is not seen as allocated and we still have a field access somewhere (which could be in dead code or code that throws a nullreference, or code that did shenanigans with Unsafe.As and actually operates on an instance of a different type).