Simplify interface handling in TypeMapStep
I can't tell if I'm deleting a bunch of cargo cults or if this is necessary for something. The tests seem to be passing.

I stumbled over this when trying to find a place where to slot in default interface method support (#427).

TypeMapStep was computing interface implementation by doing two passes:

1. For each interface & for each method on the interface, try to find a name/sig match on the current class and all the bases. If the match happens on the current class, do nothing, otherwise record the overrides/base relationship.
2. For each virtual method on the class & for each interface implemented by the class try to find a name/sig match of the virtual method on the interface. If one is found, record that info.

We were basically running two O(N*M) algorithms for something that should be possible to compute in the first pass. (We actually computed it in the first pass as described above, but we did the "do nothing" instead of recording the info we found.)

Feels like we should be able to delete the second pass.