Warnings cleanup
The most important change is to make warning codes and messages unique - that is for each code there's exactly one message.
This means adding quite a few new codes as we were sharing them rather heavily.

As I was going over these I also changes some of the messages to make them cleaner and easier to understand.

This change also adds more tests for some of the warnings.

Infra changes:
* Add the warning code parameter to unrecognized reflection call on the interface
* Add a new ExpectedWarning test attribute which validates the code, the location and optionally parts of the message
* Improved the unrecognized reflection pattern attribute to accept only parts of the message (and to accept more than one)
* Add the ability to validat the warning code for unrecognised reflection patterns