Don't generate warning 2045 due to the attribute type members referencing the type itself
If an attribute is marked for removal of instances, but the linker keeps the attribute type for other reasons (for example because it's used from a copy assembly) then there will be several other references to the attribute type from the code within the attribute type itself. These warnings are useless and should not be generated.

Other changes:
* Moved link attribute tests into their own separate folder - they don't belong to DataFlow tests
* The log verification logic now removes messages which where matched by LogContains attributes, so it's possible to validate that there are not other messages of the same kind by adding LogDoesntContain at the end