Do not call MarkInterfaceImplementation if already marked (#1396)
Fixes #1394.

The problem was that marking the interface implementation marks the custom attributes on the InterfaceImpl record which enqueues the attribute constructor. The constructor is already processed, but enqueuing a new method causes the main processing loop to run again. This will rerun the virtual method logic and cause the attribute constructor to be re-added, forever.

Normally, linker protects itself from things like this by having a "processed" state along with "marked" state for various entities. InterfaceImpls don't have a "processed" state and calling Mark will cause the same logic to run. Alternative fix would be to add a "processed" state.