Fix possible endless recursion in MarkEntireType
The method recursively calls itself to mark nested types and base types. This means that if nested type has the parent type as its base type, the method will cause stack overflow.

Fixing this by adding a hash of already visited types to break the recursion.

Added tests for the repro and similar cases.