Fix some tests
Every test .cs file must be able to resolve a type of the same name during test execution.  Otherwise it triggers a warning and if warnings ever accumulate it becomes impossible to notice when you have a test type name that doesn't match the file name

The rule of thumb to follow is, never #if out the test case type definition.  Include an ignore and and exclude any code that can't compile.

* Fix `DebuggerDisplayAttributeOnTypeCopy` only existing when `NETCOREAPP` is defined.  I just removed the define.  I don't see why it wouldn't work on .NET Framework

* Fix `DefaultInterfaceMethodCallIntoClass` only existing when `NETCOREAPP` is defined.  I most of the test is #if away when not `NETCOREAPP` app and an `[IgnoreTestCase]` is included

* Fix `GenericDefaultInterfaceMethods` only existing when `NETCOREAPP` is defined.  I most of the test is #if away when not `NETCOREAPP` app and an `[IgnoreTestCase]` is included

* Fix `SimpleDefaultInterfaceMethod` only existing when `NETCOREAPP` is defined.  I most of the test is #if away when not `NETCOREAPP` app and an `[IgnoreTestCase]` is included

* Fix `UnusedDefaultInterfaceImplementation` only existing when `NETCOREAPP` is defined.  I most of the test is #if away when not `NETCOREAPP` app and an `[IgnoreTestCase]` is included

* Fix `AttributeScopeUpdated`.  The type name did not match the file name.