[release/5.0] Unresolved attribute fixes (#1590)
When we fail to resolve a type we try and recreate it and keep the IL valid rather than fail.  We have a test that ensures an unresolved attribute on an AssemblyDefinition can be recreated.  It started failing with the sync up.   LazyMarkCustomAttributes should not skip any attributes.  This would give inconsistent behavior with references that only have attributes marked.  It was also problematic because LazyMarkCustomAttributes would call HandleUnresolvedType on the type where as ProcessLazyAttributes will call HandleUnresolvedMethod. A call to HandleUnresolvedMethod is necessary for us to recreate valid IL.

I went ahead and added null checks to other places that were missing them.

Co-authored-by: Michael Voorhees <michaelv@unity3d.com>