Fix nullref caused by unknown value to the ignoreCase parameter of Type.GetType
We have special handling for cases where ignoreCase is not constant false (since linker won't be able to correctly mark the required types). This handling was written such that it assumed the value fo the parameter was "known" that is we know at least something about it. If it's completely unknown, that's represented as null and the code crashes on it.

Added a test case for this and made the fix.